+++ src/ChartSets.js
@@ -1,8 +1,8 @@
 import React from 'react';
-import { AggregateByTimeSelectable } from './SelectableCharts';
-import { SingleDayLine, AggregateByDate } from './line';
+import { AggregateByTimeSelectable } from './charts/SelectableCharts';
+import { SingleDayLine, AggregateByDate } from './charts/line';
 import { station_direction } from './stations';
-import { BusDisclaimer } from './notes';
+import { BusDisclaimer } from './ui/notes';
 
 const dataFields = {
   traveltimes: {
