+++ src/charts/Legend.js
